requirements.txt
# Requisitos del Sistema de Gestión de Recursos Humanos

# Dependencias principales
mysql-connector-python>=8.0.0
bcrypt>=4.0.0

# Python version: 3.8+
